requirements.txt
appdirs==1.4.4
audioread==2.1.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
idna==3.3
importlib-resources==5.4.0
joblib==1.1.0
kiwisolver==1.3.1
librosa==0.9.1
llvmlite==0.36.0
matplotlib==3.3.4
numba==0.53.1
numpy==1.19.5
packaging==21.3
Pillow==8.4.0
pkg_resources==0.0.0
pooch==1.6.0
pycparser==2.21
pydub==0.25.1
pynndescent==0.5.6
pyparsing==3.0.8
python-dateutil==2.8.2
requests==2.27.1
resampy==0.2.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sklearn==0.0
SoundFile==0.10.3.post1
threadpoolctl==3.1.0
torch==1.10.2
tqdm==4.64.0
typing_extensions==4.1.1
umap-learn==0.5.3
urllib3==1.26.9
webrtcvad==2.0.10
youtube-dl==2021.12.17
zipp==3.6.0
